制作一下优化，另外加入莉莉的C结局ED

+++ Core/GlobalInstance/Players/LAPPlayerInfo.cs
@@ -1,6 +1,7 @@
 ﻿using LAP.Core.Keybind;
 using LAP.Core.NetCode;
 using Microsoft.Xna.Framework;
+using System;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
@@ -37,9 +38,9 @@ namespace LAP.Core.GlobalInstance.Players
                     // 创建一个新的网络数据包
                     ModPacket packet = Mod.GetPacket();
                     // 写入一个自定义的消息类型，以便HandlePacket能识别
-                    packet.Write((byte)LAPNetCode.MessageType.SyncMousePosition);
+                    packet.Write((int)LAPNetCode.MessageType.SyncMousePosition);
                     // 写入是哪个玩家发送的
-                    packet.Write((byte)Player.whoAmI);
+                    packet.Write(Player.whoAmI);
                     // 写入鼠标坐标
                     packet.WriteVector2(Main.MouseWorld);
                     // 发送给服务器
@@ -51,8 +52,8 @@ namespace LAP.Core.GlobalInstance.Players
                 if (Main.netMode == NetmodeID.MultiplayerClient && Main.myPlayer == Player.whoAmI)
                 {
                     ModPacket packet = Mod.GetPacket();
-                    packet.Write((byte)LAPNetCode.MessageType.SyncMouseLeft);
-                    packet.Write((byte)Player.whoAmI);
+                    packet.Write((int)LAPNetCode.MessageType.SyncMouseLeft);
+                    packet.Write(Player.whoAmI);
                     packet.Write(Main.mouseLeft);
                     packet.Send();
                 }
@@ -62,8 +63,8 @@ namespace LAP.Core.GlobalInstance.Players
                 if (Main.netMode == NetmodeID.MultiplayerClient && Main.myPlayer == Player.whoAmI)
                 {
                     ModPacket packet = Mod.GetPacket();
-                    packet.Write((byte)LAPNetCode.MessageType.SyncMouseRight);
-                    packet.Write((byte)Player.whoAmI);
+                    packet.Write((int)LAPNetCode.MessageType.SyncMouseRight);
+                    packet.Write(Player.whoAmI);
                     packet.Write(Main.mouseRight);
                     packet.Send();
                 }
@@ -73,8 +74,8 @@ namespace LAP.Core.GlobalInstance.Players
                 if (Main.netMode == NetmodeID.MultiplayerClient && Main.myPlayer == Player.whoAmI)
                 {
                     ModPacket packet = Mod.GetPacket();
-                    packet.Write((byte)LAPNetCode.MessageType.SyncWeaponSkillKey);
-                    packet.Write((byte)Player.whoAmI);
+                    packet.Write((int)LAPNetCode.MessageType.SyncWeaponSkillKey);
+                    packet.Write(Player.whoAmI);
                     packet.Write(LAPKeybind.WeaponSkillHotKey.JustPressed);
                     packet.Send();
                 }
